updated images

+++ src/Backend/Abdock.MCP/WeatherClient.Abstractions/Models/DailyWeather.cs
@@ -18,5 +18,5 @@ public record DailyWeather
     public required string ConditionText { get; init; }
     public required string ConditionIcon { get; init; }
     public required int ConditionCode { get; init; }
-    public required int UvIndex { get; init; }
+    public required double UvIndex { get; init; }
 }
